# openapi.yaml
openapi: 3.0.1
info:
  title: portal NeoHoods API
  version: "2.0"
  description: |
    # Summary
    portal NeoHoods API
servers:
  - url: https://portal.neohoods.com
    description: portal neohoods API
paths:
...
  /spaces/spaces:
    get:
      operationId: getSpaces
      summary: Retrieve all available spaces
      tags:
        - spaces-api
      parameters:
        - name: type
          in: query
          description: Filter by space type
          required: false
          schema:
            $ref: "#/components/schemas/SpaceType"
        - name: available
          in: query
          description: Filter by availability
          required: false
          schema:
            type: boolean
        - name: startDate
          in: query
          description: Filter by start date
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter by end date
          required: false
          schema:
            type: string
            format: date
        - name: search
          in: query
          description: Search query
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 12
      responses:
        "200":
          description: A paginated list of spaces
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSpaces"
  /spaces/spaces/{space-id}:
    parameters:
      - name: space-id
        in: path
        description: Space ID
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getSpace
      summary: Retrieve a specific space by ID
      tags:
        - spaces-api
      responses:
        "200":
          description: A single space
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
        "404":
          $ref: "#/components/responses/NotFound"
  /spaces/spaces/{space-id}/shared-reservations:
    parameters:
      - name: space-id
        in: path
        description: Space ID
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getSharedSpaceReservations
      summary: Retrieve reservations of spaces that share the same physical space
      description: Returns all reservations from spaces that share the physical space with this space, useful for displaying in calendars
      tags:
        - spaces-api
      parameters:
        - name: startDate
          in: query
          description: Start date for filtering reservations
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: End date for filtering reservations
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of reservations from shared spaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
        "404":
          $ref: "#/components/responses/NotFound"
  /spaces/spaces/{space-id}/availability:
    parameters:
      - name: space-id
        in: path
        description: Space ID
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getSpaceAvailability
      summary: Check availability for a space in a date range
      tags:
        - spaces-api
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Availability information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
...
  /public/stripe/webhook:
    post:
      operationId: handleStripeWebhook
      summary: Handle Stripe webhook events
      tags:
        - stripe-webhook-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Stripe event payload
      responses:
        "200":
          description: Webhook processed successfully
        "400":
          description: Invalid webhook signature
...
components:
...
  securitySchemes:
    BearerAuthOAuth:
      description: Bearer Token.
      type: http
      scheme: bearer
...
  schemas:
...
    SpaceType:
      type: string
      enum:
        - GUEST_ROOM
        - COMMON_ROOM
        - COWORKING
        - PARKING
      description: Type of space available for reservation
    SpaceStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - MAINTENANCE
      description: Current status of the space
    Space:
      type: object
      required:
        - id
        - name
        - type
        - status
        - description
        - instructions
        - pricing
        - rules
        - images
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Chambre d'amis - Appartement 2A"
        type:
          $ref: "#/components/schemas/SpaceType"
        status:
          $ref: "#/components/schemas/SpaceStatus"
        description:
          type: string
          example: "Chambre d'amis de 15m² avec lit double, bureau et placard"
        instructions:
          type: string
          example: "Code d'accès envoyé par email. Nettoyage obligatoire avant départ."
        pricing:
          $ref: "#/components/schemas/SpacePricing"
        rules:
          $ref: "#/components/schemas/SpaceRules"
        images:
          type: array
          items:
            $ref: "#/components/schemas/SpaceImage"
        quota:
          $ref: "#/components/schemas/QuotaInfo"
        digitalLockId:
          type: string
          format: uuid
          description: "ID of associated digital lock"
        accessCodeEnabled:
          type: boolean
          default: true
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SpacePricing:
      type: object
      required:
        - ownerPrice
        - tenantPrice
        - currency
      properties:
        ownerPrice:
          type: number
          format: float
          example: 15.0
          description: "Daily price for owners (propriétaires)"
        tenantPrice:
          type: number
          format: float
          example: 20.0
          description: "Daily price for tenants (locataires)"
        currency:
          type: string
          example: "EUR"
        cleaningFee:
          type: number
          format: float
          example: 25.0
        deposit:
          type: number
          format: float
          example: 50.0
...
    SpaceRules:
      type: object
      required:
        - minDurationDays
        - maxDurationDays
        - maxReservationsPerYear
        - allowedDays
        - allowedHours
      properties:
        minDurationDays:
          type: integer
          minimum: 1
          example: 3
        maxDurationDays:
          type: integer
          minimum: 1
          example: 7
        maxReservationsPerYear:
          type: integer
          minimum: 1
          example: 4
        allowedDays:
          type: array
          items:
            type: string
            enum:
              [
                "MONDAY",
                "TUESDAY",
                "WEDNESDAY",
                "THURSDAY",
                "FRIDAY",
                "SATURDAY",
                "SUNDAY",
              ]
          example: ["WEDNESDAY", "SATURDAY", "SUNDAY"]
        allowedHours:
          $ref: "#/components/schemas/TimeRange"
        cleaningDays:
          type: array
          items:
            type: string
            enum:
              [
                "MONDAY",
                "TUESDAY",
                "WEDNESDAY",
                "THURSDAY",
                "FRIDAY",
                "SATURDAY",
                "SUNDAY",
              ]
          example: ["MONDAY", "FRIDAY"]
        requiresApartmentAccess:
          type: boolean
          example: true
        shareSpaceWith:
          type: array
          items:
            type: string
            format: uuid
          description: "Liste des UUIDs des espaces qui partagent le même espace physique. La bidirectionnalité est gérée automatiquement par le backend."
          example: ["550e8400-e29b-41d4-a716-446655440001"]
    TimeRange:
      type: object
      required:
        - start
        - end
      properties:
        start:
          type: string
          pattern: "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
          example: "08:00"
        end:
          type: string
          pattern: "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
          example: "20:00"
    QuotaInfo:
      type: object
      required:
        - max
        - period
      properties:
        max:
          type: integer
          minimum: 1
          example: 4
        period:
          type: string
          enum: ["YEAR", "MONTH", "WEEK"]
          example: "YEAR"
        resetDate:
          type: string
          format: date
          example: "2024-01-01"
    ReservationStatus:
      type: string
      enum:
        - PENDING_PAYMENT
        - PAYMENT_FAILED
        - EXPIRED
        - CONFIRMED
        - ACTIVE
        - COMPLETED
        - CANCELLED
        - REFUNDED
      description: Current status of the reservation
    Reservation:
      type: object
      required:
        - id
        - spaceId
        - userId
        - startDate
        - endDate
        - status
        - totalPrice
        - currency
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        spaceId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        space:
          $ref: "#/components/schemas/Space"
        user:
          $ref: "#/components/schemas/User"
        startDate:
          type: string
          format: date
          example: "2024-02-15"
        endDate:
          type: string
          format: date
          example: "2024-02-18"
        status:
          $ref: "#/components/schemas/ReservationStatus"
        totalPrice:
          type: number
          format: float
          example: 60.0
        currency:
          type: string
          example: "EUR"
        cleaningFee:
          type: number
          format: float
          example: 25.0
        deposit:
          type: number
          format: float
          example: 50.0
        platformFeeAmount:
          type: number
          format: float
          example: 1.2
          description: "Platform fee amount (percentage-based)"
        platformFixedFeeAmount:
          type: number
          format: float
          example: 0.25
          description: "Fixed platform fee amount per transaction"
        stripeSessionId:
          type: string
          example: "cs_test_123456789"
        stripePaymentIntentId:
          type: string
          example: "pi_123456789"
        accessCode:
          $ref: "#/components/schemas/AccessCode"
        feedback:
          $ref: "#/components/schemas/ReservationFeedback"
        paymentExpiresAt:
          type: string
          format: date-time
          description: "When the payment expires (15 minutes after creation)"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AvailabilityResponse:
      type: object
      required:
        - available
        - conflicts
        - quota
      properties:
        available:
          type: boolean
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/ReservationConflict"
        quota:
          $ref: "#/components/schemas/QuotaInfo"
        price:
          type: number
          format: float
        currency:
          type: string
    ReservationConflict:
      type: object
      required:
        - type
        - message
      properties:
        type:
          type: string
          enum:
            [
              "SAME_DAY_BLOCK",
              "QUOTA_EXCEEDED",
              "CLEANING_DAY",
              "VACATION_PERIOD",
            ]
        message:
          type: string
        conflictingReservationId:
          type: string
          format: uuid
    AccessCode:
      type: object
      required:
        - code
        - generatedAt
        - expiresAt
      properties:
        code:
          type: string
          example: "ABC123"
        generatedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isActive:
          type: boolean
          example: true
    ReservationFeedback:
      type: object
      required:
        - rating
        - comment
        - submittedAt
      properties:
        id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        comment:
          type: string
          maxLength: 1000
        cleanliness:
          type: integer
          minimum: 1
          maximum: 5
        communication:
          type: integer
          minimum: 1
          maximum: 5
        value:
          type: integer
          minimum: 1
          maximum: 5
        submittedAt:
          type: string
          format: date-time
...
    PaginatedSpaces:
      type: object
      required:
        - content
        - totalElements
        - totalPages
        - size
        - number
        - first
        - last
        - numberOfElements
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Space"
        totalElements:
          type: long
          description: Total number of elements across all pages
        totalPages:
          type: integer
          description: Total number of pages
        size:
          type: integer
          description: Number of elements per page
        number:
          type: integer
          description: Current page number (0-based)
        first:
          type: boolean
          description: Whether this is the first page
        last:
          type: boolean
          description: Whether this is the last page
        numberOfElements:
          type: integer
          description: Number of elements in current page
...
    User:
      type: object
      required:
        - id
        - username
        - firstName
        - lastName
        - email
        - isEmailVerified
        - disabled
        - type
        - properties
        - streetAddress
        - city
        - postalCode
        - country
      properties:
        id:
          type: string
          format: uuid
        disabled:
          type: boolean
        isEmailVerified:
          type: boolean
        username:
          type: string
        avatarUrl:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        flatNumber:
          type: string
          example: A701
        roles:
          type: array
          items:
            type: string
            example: "admin"
        streetAddress:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        preferredLanguage:
          type: string
          example: "en"
        profileSharingConsent:
          type: boolean
          description: "User consent for sharing profile information in the community directory"
          default: false
        createdAt:
          type: string
          format: date-time
          description: "User registration date"
        type:
          $ref: "#/components/schemas/UserType"
        properties:
          type: array
          items:
            $ref: "#/components/schemas/Property"
    UserType:
      type: string
      enum:
        - ADMIN
        - OWNER
        - LANDLORD
        - TENANT
        - SYNDIC
        - EXTERNAL
        - CONTRACTOR
        - COMMERCIAL_PROPERTY_OWNER
        - GUEST
      example: "OWNER"
    Property:
      type: object
      required:
        - type
        - name
      properties:
        type:
          $ref: "#/components/schemas/PropertyType"
        name:
          type: string
          example: "Apartment 3A"
    PropertyType:
      type: string
      enum:
        - APARTMENT
        - GARAGE
        - PARKING
        - COMMERCIAL
        - OTHER
      example: "APARTMENT"
...
    CodedError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: A unique error code that identifies the type of error
          example: AUTH001
        message:
          type: string
          description: A human-readable error message
          example: The specified user could not be found
        traceId:
          type: string
          description: A unique identifier for tracing this error instance
          example: 550e8400-e29b-41d4-a716-446655440000
        documentationUrl:
          type: string
          description: URL to the documentation about this error
          example: https://docs.portal.neohoods.com/errors/AUTH001
        variables:
          type: object
          additionalProperties: true
          description: Additional context variables for the error
          example:
            field: "email"
            value: "invalid@email"
...
    SpaceImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        spaceId:
          type: string
          format: uuid
        url:
          type: string
          format: uri
          description: External image URL (optional)
        mimeType:
          type: string
          description: MIME type for images stored in database
        fileName:
          type: string
          description: Original filename
        fileSize:
          type: integer
          description: File size in bytes
        altText:
          type: string
          description: Alt text for accessibility
        isPrimary:
          type: boolean
          description: Whether this is the primary image
        orderIndex:
          type: integer
          description: Display order
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
...
    NotFound:
      description: Not Found - The requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CodedError"
          example:
            code: RES001
            message: The requested resource was not found
...